requirements.txt
 Django==1.6
 psycopg2
 eav-django
 django-autoslug
